+++ app/models/bids.py
@@ -15,40 +15,38 @@
 
 from datetime import datetime
 
+from sqlalchemy import JSON
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import String
 from sqlalchemy.dialects.postgresql import TIMESTAMP
 
 from app.config import ConfigClass
-from app.core.db import DBModel
+from app.models import DBModel
 
 
-class DatasetVersion(DBModel):
-    __tablename__ = 'dataset_version'
+class BIDSResult(DBModel):
+    __tablename__ = 'bids_results'
     __table_args__ = {'schema': ConfigClass.RDS_SCHEMA_DEFAULT}
     id = Column(Integer, primary_key=True)
-    dataset_code = Column(String())
     dataset_geid = Column(String())
-    version = Column(String())
-    created_by = Column(String())
-    created_at = Column(TIMESTAMP(timezone=True), default=datetime.utcnow, nullable=False)
-    location = Column(String())
-    notes = Column(String())
+    created_time = Column(TIMESTAMP(timezone=True), default=datetime.utcnow, nullable=False)
+    updated_time = Column(TIMESTAMP(timezone=True), default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False)
+    validate_output = Column(JSON())
 
-    def __init__(self, dataset_code, dataset_geid, version, created_by, location, notes):
-        self.dataset_code = dataset_code
+    def __init__(self, dataset_geid, validate_output):
         self.dataset_geid = dataset_geid
-        self.version = version
-        self.created_by = created_by
-        self.location = location
-        self.notes = notes
+        self.validate_output = validate_output
+        if self.created_time:
+            self.created_time = self.created_time
 
     def to_dict(self):
         result = {}
-        for field in ['id', 'dataset_code', 'dataset_geid', 'version', 'created_by', 'created_at', 'location', 'notes']:
-            if field == 'created_at':
+        for field in ['id', 'dataset_geid', 'created_time', 'updated_time', 'validate_output']:
+            if field == 'created_time' or field == 'updated_time':
                 result[field] = str(getattr(self, field).isoformat()[:-3] + 'Z')
+            elif field == 'validate_output':
+                result[field] = getattr(self, field)
             else:
                 result[field] = str(getattr(self, field))
         return result
