+++ app/routers/v1/api_schema/api_schema_template.py
@@ -21,7 +21,7 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from app.commons.logger_services.logger_factory_service import SrvLoggerFactory
 from app.core.db import get_db_session
-from app.models.schema_sql import DatasetSchemaTemplate
+from app.models.schema import DatasetSchemaTemplate
 from app.resources.error_handler import catch_internal
 from app.schemas.base import APIResponse
 from app.schemas.base import EAPIResponseCode
