+++ app/routers/v1/api_version/publish_version.py
@@ -26,8 +26,8 @@ from redis import Redis
 from app.commons.logger_services.logger_factory_service import SrvLoggerFactory
 from app.commons.service_connection.minio_client import Minio_Client
 from app.config import ConfigClass
-from app.models.schema_sql import DatasetSchema
-from app.models.version_sql import DatasetVersion
+from app.models.schema import DatasetSchema
+from app.models.version import DatasetVersion
 from app.resources.locks import recursive_lock_publish
 from app.resources.locks import unlock_resource
 from app.resources.neo4j_helper import get_children_nodes
