+++ tests/api_dataset_restful/conftest.py
@@ -4,7 +4,7 @@ import pytest
 @pytest.fixture
 def schema_essential_template(db_session):
     from app.config import ConfigClass
-    from app.models.schema_sql import DatasetSchemaTemplate
+    from app.models.schema import DatasetSchemaTemplate
 
     dataset_geid = '5baeb6a1-559b-4483-aadf-ef60519584f3'
     schema_template_geid = 'ef4eb37d-6d81-46a7-a9d9-db71bf44edc7'
@@ -28,7 +28,7 @@ def schema_essential_template(db_session):
 
 @pytest.fixture
 def bids_results(db_session):
-    from app.models.bids_sql import BIDSResult
+    from app.models.bids import BIDSResult
 
     dataset_geid = '5baeb6a1-559b-4483-aadf-ef60519584f3'
